+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes are documented here.
 
 ## 1.1.5 - 2026-08-04
 
-- Render cover-letter names in uppercase to match resume headers.
+- Add descriptive labels to cover-letter contact details
 
 ## 1.1.4 - 2026-08-04